requirements.txt
pandas
numpy
yfinance
alpaca-py
backtesting.py
scikit-learn
tensorflow
torch
xgboost
matplotlib
pandas-market-calendars
pyyaml
hmmlearn
requests
websocket-client
python-dotenv
